blip_requirements.txt
timm==0.4.12
transformers==4.17.0
fairscale==0.4.4
pycocoevalcap
torch
torchvision
Pillow